refactor(tests): split fake repository documents
